--story=856753609 1）邮件支持附件 2）codecc external resource包调整 3）codecc、红线bug小调整
