+++ src/backend/ci/core/notify/api-notify/src/main/kotlin/com/tencent/devops/notify/pojo/EmailNotifyMessage.kt
@@ -55,6 +55,8 @@ open class EmailNotifyMessage : BaseMessage() {
     var priority: EnumNotifyPriority = EnumNotifyPriority.HIGH
     @ApiModelProperty("通知来源", allowableValues = "0,1", dataType = "int")
     var source: EnumNotifySource = EnumNotifySource.BUSINESS_LOGIC
+    @ApiModelProperty("codecc邮件附件内容")
+    var codeccAttachFileContent: Map<String, String>? = mapOf()
 
     fun addReceiver(receiver: String) {
         receivers.add(receiver)
