+++ src/backend/ci/core/plugin/codecc-plugin/api-codecc/src/main/kotlin/com/tencent/devops/plugin/api/ServiceCodeccElementResource.kt
@@ -24,50 +24,34 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.plugin.codecc.api
+package com.tencent.devops.plugin.api
 
-import com.tencent.devops.common.api.auth.AUTH_HEADER_USER_ID
+import com.tencent.devops.common.api.annotation.ServiceInterface
 import com.tencent.devops.common.api.pojo.Result
-import com.tencent.devops.plugin.codecc.pojo.CodeccCallback
+import com.tencent.devops.plugin.codecc.pojo.CodeccElementData
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
-import io.swagger.annotations.ApiParam
 import javax.ws.rs.Consumes
 import javax.ws.rs.GET
-import javax.ws.rs.HeaderParam
 import javax.ws.rs.Path
 import javax.ws.rs.PathParam
 import javax.ws.rs.Produces
-import javax.ws.rs.QueryParam
 import javax.ws.rs.core.MediaType
 
-@Api(tags = ["USER_CODECC"], description = "用户-codecc相关")
-@Path("/user/codecc")
+@Api(tags = ["SERVICE_CODECC_ELEMENT"])
+@Path("/service/codecc/element")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-interface UserCodeccResource {
+@ServiceInterface("plugin") // 指明接入到哪个微服务
+interface ServiceCodeccElementResource {
 
-    @ApiOperation("获取codecc报告")
+    @ApiOperation("提供根据流水线构建id查询构建号（页面上展示用的）、构建时间、构建人等信息件")
     @GET
-    @Path("/report/{buildId}")
-    fun getCodeccReport(
-        @ApiParam("构建ID", required = true)
-        @PathParam(value = "buildId")
-        buildId: String
-    ): Result<CodeccCallback?>
-
-    @ApiOperation("获取codecc规则集")
-    @GET
-    @Path("/ruleSet/project/{projectId}/codeccRuleSet")
-    fun getCodeccRuleSet(
-        @ApiParam("项目ID", required = true)
-        @PathParam(value = "projectId")
+    @Path("/project/{projectId}/pipeline/{pipelineId}")
+    fun get(
+        @PathParam("projectId")
         projectId: String,
-        @ApiParam("用户ID", required = true)
-        @HeaderParam(AUTH_HEADER_USER_ID)
-        userId: String,
-        @ApiParam("工具名称", required = true)
-        @QueryParam(value = "toolName")
-        toolName: String
-    ): Result<Map<String, Any>>
+        @PathParam("pipelineId")
+        pipelineId: String
+    ): Result<CodeccElementData?>
 }
