+++ src/backend/ci/core/plugin/codecc-plugin/api-codecc/src/main/kotlin/com/tencent/devops/plugin/api/UserCodeccResource.kt
@@ -24,13 +24,11 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.plugin.codecc.api
+package com.tencent.devops.plugin.api
 
-import com.tencent.devops.common.api.annotation.ServiceInterface
 import com.tencent.devops.common.api.auth.AUTH_HEADER_USER_ID
-import com.tencent.devops.common.api.auth.AUTH_HEADER_USER_ID_DEFAULT_VALUE
 import com.tencent.devops.common.api.pojo.Result
-import com.tencent.devops.plugin.codecc.pojo.coverity.CodeccReport
+import com.tencent.devops.plugin.codecc.pojo.CodeccCallback
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiParam
@@ -40,28 +38,36 @@ import javax.ws.rs.HeaderParam
 import javax.ws.rs.Path
 import javax.ws.rs.PathParam
 import javax.ws.rs.Produces
+import javax.ws.rs.QueryParam
 import javax.ws.rs.core.MediaType
 
-@Api(tags = ["USER_BUILD"], description = "用户-构建资源-扩展")
-@Path("/user/builds")
+@Api(tags = ["USER_CODECC"], description = "用户-codecc相关")
+@Path("/user/codecc")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-@ServiceInterface("plugin") // 指明接入到哪个微服务
-interface UserBuildCodeccResource {
+interface UserCodeccResource {
 
-    @ApiOperation("获取CodeCC报告")
+    @ApiOperation("获取codecc报告")
     @GET
-    // @Path("/project/{projectId}/pipeline/{pipelineId}/codeccReport")
-    @Path("/{projectId}/{pipelineId}/codeccReport")
+    @Path("/report/{buildId}")
     fun getCodeccReport(
-        @ApiParam(value = "用户ID", required = true, defaultValue = AUTH_HEADER_USER_ID_DEFAULT_VALUE)
+        @ApiParam("构建ID", required = true)
+        @PathParam(value = "buildId")
+        buildId: String
+    ): Result<CodeccCallback?>
+
+    @ApiOperation("获取codecc规则集")
+    @GET
+    @Path("/ruleSet/project/{projectId}/codeccRuleSet")
+    fun getCodeccRuleSet(
+        @ApiParam("项目ID", required = true)
+        @PathParam(value = "projectId")
+        projectId: String,
+        @ApiParam("用户ID", required = true)
         @HeaderParam(AUTH_HEADER_USER_ID)
         userId: String,
-        @ApiParam("项目ID", required = true)
-        @PathParam("projectId")
-        projectId: String,
-        @ApiParam("流水线ID", required = true)
-        @PathParam("pipelineId")
-        pipelineId: String
-    ): Result<CodeccReport>
+        @ApiParam("工具名称", required = true)
+        @QueryParam(value = "toolName")
+        toolName: String
+    ): Result<Map<String, Any>>
 }
