+++ src/backend/ci/core/plugin/codecc-plugin/biz-codecc/src/main/kotlin/com/tencent/devops/plugin/codecc/resources/ExternalCodeccResourceImpl.kt
@@ -28,7 +28,7 @@ package com.tencent.devops.plugin.codecc.resources
 
 import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.web.RestResource
-import com.tencent.devops.plugin.codecc.api.ExternalCodeccResource
+import com.tencent.devops.plugin.api.ExternalCodeccResource
 import com.tencent.devops.plugin.codecc.pojo.CodeccCallback
 import com.tencent.devops.plugin.codecc.service.CodeccService
 import org.springframework.beans.factory.annotation.Autowired
