+++ src/backend/ci/core/plugin/codecc-plugin/biz-codecc/src/main/kotlin/com/tencent/devops/plugin/codecc/resources/UserBuildCodeccResourceImpl.kt
@@ -29,7 +29,7 @@ package com.tencent.devops.plugin.codecc.resources
 import com.tencent.devops.common.api.exception.ParamBlankException
 import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.web.RestResource
-import com.tencent.devops.plugin.codecc.api.UserBuildCodeccResource
+import com.tencent.devops.plugin.api.UserBuildCodeccResource
 import com.tencent.devops.plugin.codecc.pojo.coverity.CodeccReport
 import com.tencent.devops.plugin.codecc.service.PipelineCodeccService
 import org.springframework.beans.factory.annotation.Autowired
